Use valid but empty spaces on env construction
Use a valid but empty instance of gym.spaces.Space as sentinel when
constructing the VeinsGym env object (instead of None).

Newer versions (0.24.1+) of Gym contain an env checker that would fail
otherwise if the observation and action spaces are not valid Space-s.
VeinsGym only sets (or overrides) the action and observation spaces on
the reset() call, because only then the connection to the Veins instance
is established and the space definitions can be obtained.

Also added an assert to ensure spaces are valid when parsing requests.

Fixes #8