WIP: extend interface of reset() as required for future gym versions
As current warnings tell, reset() will be required to accept extra
arguments in future Gym versions: seed, return_info, and options.
This is a try to minimally implement that.

I wanted to use an explicit sentinel for seed to identify whether None
was passed explicitly. Calling VeinsGym.reset(None) does change the seed
(randomly), so there needs to be an opt-out of seed changes in reset().
At least that is my current understanding/oppinon.

This needs further research, thinking, and implementaiton.